Refactor: RL1M-304 신규 웹소켓 폴더 구조 개선
